+++ src/main/java/org/baeldung/Application.java
@@ -1,4 +1,4 @@
-package org.baeldung.spring;
+package org.baeldung;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
